chore: remove dead code, add CI workflow, fix timestamp type mismatch
- Delete unused src/consumer/consumer.py (superseded by spark_streaming.py)
- Add GitHub Actions workflow running ruff check, ruff format, and mypy on PRs and main
- Fix ruff/mypy violations so the lint suite passes cleanly
- Align event_timestamp with the Avro long type: BIGINT in Cassandra DDL, int(time.time()) in the producer
- Bump pre-commit hook versions to match pinned CI tool versions
- Add CI status badge to README

+++ src/cassandra/setup.sql
@@ -7,7 +7,7 @@ WITH replication = {
 USE events;
 
 CREATE TABLE user_actions (
-    event_timestamp INT,
+    event_timestamp BIGINT,
     event_id INT,
     event_type TEXT,
     device_type TEXT,
